feat: refactor providers and add together provider
