+++ requirements.txt
@@ -1,6 +1,12 @@
 # This file was autogenerated by uv via the following command:
 #    uv export --no-hashes --output-file=requirements.txt
 -e .
+aiohappyeyeballs==2.6.1
+    # via aiohttp
+aiohttp==3.12.15
+    # via together
+aiosignal==1.4.0
+    # via aiohttp
 annotated-types==0.7.0
     # via pydantic
 anyio==4.9.0
@@ -12,16 +18,23 @@ appnope==0.1.4 ; sys_platform == 'darwin'
     # via ipykernel
 asttokens==3.0.0
     # via stack-data
+attrs==25.3.0
+    # via aiohttp
 certifi==2025.6.15
     # via
     #   httpcore
     #   httpx
+    #   requests
 cffi==1.17.1 ; implementation_name == 'pypy'
     # via pyzmq
 cfgv==3.4.0
     # via pre-commit
+charset-normalizer==3.4.3
+    # via requests
 click==8.2.1
-    # via typer
+    # via
+    #   together
+    #   typer
 colorama==0.4.6 ; sys_platform == 'win32'
     # via
     #   click
@@ -42,11 +55,19 @@ distro==1.9.0
     #   groq
     #   openai
 eval-type-backport==0.2.2
-    # via mistralai
+    # via
+    #   mistralai
+    #   together
 executing==2.2.0
     # via stack-data
 filelock==3.18.0
-    # via virtualenv
+    # via
+    #   together
+    #   virtualenv
+frozenlist==1.7.0
+    # via
+    #   aiohttp
+    #   aiosignal
 greenlet==3.2.3 ; (python_full_version < '3.14' and platform_machine == 'AMD64') or (python_full_version < '3.14' and platform_machine == 'WIN32') or (python_full_version < '3.14' and platform_machine == 'aarch64') or (python_full_version < '3.14' and platform_machine == 'amd64') or (python_full_version < '3.14' and platform_machine == 'ppc64le') or (python_full_version < '3.14' and platform_machine == 'win32') or (python_full_version < '3.14' and platform_machine == 'x86_64')
     # via sqlalchemy
 groq==0.31.0
@@ -67,6 +88,8 @@ idna==3.10
     # via
     #   anyio
     #   httpx
+    #   requests
+    #   yarl
 iniconfig==2.1.0
     # via pytest
 ipykernel==6.29.5
@@ -84,7 +107,7 @@ jupyter-core==5.8.1
     # via
     #   ipykernel
     #   jupyter-client
-markdown-it-py==3.0.0
+markdown-it-py==4.0.0
     # via rich
 matplotlib-inline==0.1.7
     # via
@@ -94,10 +117,16 @@ mdurl==0.1.2
     # via markdown-it-py
 mistralai==1.9.2
     # via batchling
+multidict==6.6.4
+    # via
+    #   aiohttp
+    #   yarl
 nest-asyncio==1.6.0
     # via ipykernel
 nodeenv==1.9.1
     # via pre-commit
+numpy==2.3.2
+    # via together
 openai==1.93.0
     # via batchling
 packaging==25.0
@@ -108,6 +137,8 @@ parso==0.8.4
     # via jedi
 pexpect==4.9.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'
     # via ipython
+pillow==11.3.0
+    # via together
 platformdirs==4.3.8
     # via
     #   batchling
@@ -118,6 +149,10 @@ pluggy==1.6.0
 pre-commit==4.2.0
 prompt-toolkit==3.0.51
     # via ipython
+propcache==0.3.2
+    # via
+    #   aiohttp
+    #   yarl
 psutil==7.0.0
     # via ipykernel
 ptyprocess==0.7.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'
@@ -131,6 +166,7 @@ pydantic==2.11.7
     #   groq
     #   mistralai
     #   openai
+    #   together
 pydantic-core==2.33.2
     # via pydantic
 pygments==2.19.2
@@ -153,8 +189,12 @@ pyzmq==27.0.0
     # via
     #   ipykernel
     #   jupyter-client
-rich==14.0.0
-    # via typer
+requests==2.32.4
+    # via together
+rich==14.1.0
+    # via
+    #   together
+    #   typer
 ruff==0.12.4
 shellingham==1.5.4
     # via typer
@@ -169,12 +209,18 @@ sqlalchemy==2.0.41
     # via batchling
 stack-data==0.6.3
     # via ipython
+tabulate==0.9.0
+    # via together
+together==1.5.25
+    # via batchling
 tornado==6.5.1
     # via
     #   ipykernel
     #   jupyter-client
 tqdm==4.67.1
-    # via openai
+    # via
+    #   openai
+    #   together
 traitlets==5.14.3
     # via
     #   comm
@@ -183,10 +229,13 @@ traitlets==5.14.3
     #   jupyter-client
     #   jupyter-core
     #   matplotlib-inline
-typer==0.16.0
-    # via batchling
+typer==0.15.3
+    # via
+    #   batchling
+    #   together
 typing-extensions==4.14.0
     # via
+    #   aiosignal
     #   anyio
     #   groq
     #   openai
@@ -199,7 +248,11 @@ typing-inspection==0.4.1
     # via
     #   mistralai
     #   pydantic
+urllib3==2.5.0
+    # via requests
 virtualenv==20.31.2
     # via pre-commit
 wcwidth==0.2.13
     # via prompt-toolkit
+yarl==1.20.1
+    # via aiohttp
