+++ src/batchling/__init__.py
@@ -1,4 +1,13 @@
-from .experiment import OpenAIExperiment as OpenAIExperiment
 from .experiment_manager import ExperimentManager as ExperimentManager
+from .providers.groq import GroqExperiment as GroqExperiment
+from .providers.mistral import MistralExperiment as MistralExperiment
+from .providers.openai import OpenAIExperiment as OpenAIExperiment
+from .providers.together import TogetherExperiment as TogetherExperiment
 
-__all__ = ["OpenAIExperiment", "ExperimentManager"]
+__all__ = [
+    "OpenAIExperiment",
+    "MistralExperiment",
+    "GroqExperiment",
+    "TogetherExperiment",
+    "ExperimentManager",
+]
