+++ src/batchling/cls_utils.py
@@ -1,11 +1,10 @@
 from typing import Type
 
-from batchling.experiment import (
-    Experiment,
-    GroqExperiment,
-    MistralExperiment,
-    OpenAIExperiment,
-)
+from batchling.experiment import Experiment
+from batchling.providers.groq import GroqExperiment
+from batchling.providers.mistral import MistralExperiment
+from batchling.providers.openai import OpenAIExperiment
+from batchling.providers.together import TogetherExperiment
 
 
 def get_experiment_cls_from_provider(provider: str | None = None) -> Type[Experiment]:
@@ -13,5 +12,7 @@ def get_experiment_cls_from_provider(provider: str | None = None) -> Type[Experi
         return MistralExperiment
     elif provider == "groq":
         return GroqExperiment
+    elif provider == "together":
+        return TogetherExperiment
     else:
         return OpenAIExperiment
