+++ src/batchling/experiment.py
@@ -4,14 +4,9 @@ from abc import ABC, abstractmethod
 from datetime import datetime
 from functools import cached_property
 
-from groq import Groq
-from groq.resources.batches import BatchRetrieveResponse
-from groq.resources.files import BinaryAPIResponse, FileInfoResponse
-from httpx import Response
 from mistralai import Mistral
 from mistralai.models import BatchJobOut, RetrieveFileOut
 from openai import OpenAI
-from openai._legacy_response import HttpxBinaryResponseContent
 from openai.types.batch import Batch
 from openai.types.file_object import FileObject
 from pydantic import (
@@ -28,10 +23,6 @@ from batchling.db.session import get_db, init_db
 from batchling.file_utils import write_jsonl_file
 from batchling.request import (
     Body,
-    GroqBody,
-    GroqRequest,
-    MistralBody,
-    MistralRequest,
     OpenAIBody,
     OpenAIRequest,
     Request,
@@ -152,7 +143,7 @@ class Experiment(BaseModel, ABC):
         write_jsonl_file(file_path=self.input_file_path, data=batch_requests)
 
     @abstractmethod
-    def get_provider_results(self):
+    def get_provider_results(self) -> t.Any:
         pass
 
     @abstractmethod
@@ -238,12 +229,7 @@ class Experiment(BaseModel, ABC):
         with get_db() as db:
             update_experiment(db=db, id=self.id, updated_at=datetime.now())
 
-    def get_results(self) -> HttpxBinaryResponseContent:
-        """Get the results of the experiment
-
-        Returns:
-            HttpxBinaryResponseContent: The results
-        """
+    def get_results(self):
         self.raise_not_in_completed_status()
         return self.get_provider_results()
 
@@ -319,297 +305,3 @@ class Experiment(BaseModel, ABC):
         if db_experiment is None:
             raise ValueError(f"Experiment with id: {self.id} not found")
         return self.model_validate(db_experiment)
-
-
-class OpenAIExperiment(Experiment):
-    body_cls: t.Type[OpenAIBody] = Field(
-        default=OpenAIBody, description="body class to use", init=False
-    )
-    request_cls: t.Type[OpenAIRequest] = Field(
-        default=OpenAIRequest, description="request class to use", init=False
-    )
-
-    @computed_field(repr=False)
-    @cached_property
-    def client(self) -> OpenAI:
-        """Get the client
-
-        Returns:
-            OpenAI: The client
-        """
-        return OpenAI(api_key=os.getenv(self.api_key_name), base_url=self.base_url)
-
-    def retrieve_provider_file(self):
-        return self.client.files.retrieve(self.input_file_id)
-
-    def retrieve_provider_batch(self):
-        return self.client.batches.retrieve(self.batch_id)
-
-    @computed_field
-    @property
-    def input_file(self) -> FileObject | None:
-        if self.input_file_id is None:
-            return None
-        return self.retrieve_provider_file()
-
-    @computed_field
-    @property
-    def batch(self) -> Batch | None:
-        if self.batch_id is None:
-            return None
-        return self.retrieve_provider_batch()
-
-    @computed_field
-    @property
-    def status(
-        self,
-    ) -> t.Literal[
-        "setup",
-        "created",
-        "validating",
-        "failed",
-        "in_progress",
-        "finalizing",
-        "completed",
-        "expired",
-        "cancelling",
-        "cancelled",
-    ]:
-        if self.batch_id is None:
-            if self.is_setup:
-                return "setup"
-            return "created"
-        return self.batch.status
-
-    def create_provider_file(self) -> str:
-        return self.client.files.create(file=open(self.input_file_path, "rb"), purpose="batch").id
-
-    def delete_provider_file(self, file_id: str):
-        self.client.files.delete(file_id=file_id)
-
-    def create_provider_batch(self) -> str:
-        return self.client.batches.create(
-            input_file_id=self.input_file_id,
-            endpoint=self.endpoint,
-            completion_window="24h",
-            metadata={"description": self.description},
-        ).id
-
-    def raise_not_in_running_status(self):
-        if self.status != "running":
-            raise ValueError(f"Experiment in status {self.status} is not in running status")
-
-    def raise_not_in_completed_status(self):
-        if self.status != "completed":
-            raise ValueError(f"Experiment in status {self.status} is not in completed status")
-
-    def cancel_provider_batch(self):
-        self.client.batches.cancel(self.batch_id)
-
-    def delete_provider_batch(self):
-        if self.batch.status == "in_progress":
-            self.cancel_provider_batch()
-        elif self.batch.status == "completed" and self.batch.output_file_id:
-            self.delete_provider_file(file_id=self.batch.output_file_id)
-
-    def get_provider_results(self) -> HttpxBinaryResponseContent:
-        return self.client.files.content(file_id=self.batch.output_file_id)
-
-
-class MistralExperiment(Experiment):
-    body_cls: t.Type[MistralBody] = Field(
-        default=MistralBody, description="body class to use", init=False
-    )
-    request_cls: t.Type[MistralRequest] = Field(
-        default=MistralRequest, description="request class to use", init=False
-    )
-
-    @computed_field(repr=False)
-    @cached_property
-    def client(self) -> Mistral:
-        """Get the client
-
-        Returns:
-            Mistral: The client
-        """
-        return Mistral(api_key=os.getenv(self.api_key_name))
-
-    def retrieve_provider_file(self):
-        return self.client.files.retrieve(file_id=self.input_file_id)
-
-    def retrieve_provider_batch(self):
-        return self.client.batch.jobs.get(job_id=self.batch_id)
-
-    @computed_field
-    @property
-    def input_file(self) -> RetrieveFileOut | None:
-        if self.input_file_id is None:
-            return None
-        return self.retrieve_provider_file()
-
-    @computed_field
-    @property
-    def batch(self) -> BatchJobOut | None:
-        if self.batch_id is None:
-            return None
-        return self.retrieve_provider_batch()
-
-    @computed_field
-    @property
-    def status(
-        self,
-    ) -> t.Literal[
-        "setup",
-        "created",
-        "QUEUED",
-        "RUNNING",
-        "SUCCESS",
-        "FAILED",
-        "TIMEOUT_EXCEEDED",
-        "CANCELLATION_REQUESTED",
-        "CANCELLED",
-    ]:
-        if self.batch_id is None:
-            if self.is_setup:
-                return "setup"
-            return "created"
-        return self.batch.status
-
-    def create_provider_file(self) -> str:
-        return self.client.files.upload(
-            file={
-                "file_name": self.input_file_path.split("/")[-1],
-                "content": open(self.input_file_path, "rb"),
-            },
-            purpose="batch",
-        ).id
-
-    def delete_provider_file(self, file_id: str):
-        self.client.files.delete(file_id=file_id)
-
-    def create_provider_batch(self) -> str:
-        return self.client.batch.jobs.create(
-            input_files=[self.input_file_id],
-            endpoint=self.endpoint,
-            model=self.model,
-            metadata={"description": self.description},
-        ).id
-
-    def raise_not_in_running_status(self):
-        if self.status not in ["QUEUED", "RUNNING"]:
-            raise ValueError(
-                f"Experiment in status {self.status} is not in QUEUED or RUNNING status"
-            )
-
-    def raise_not_in_completed_status(self):
-        if self.status != "SUCCESS":
-            raise ValueError(f"Experiment in status {self.status} is not in SUCCESS status")
-
-    def cancel_provider_batch(self):
-        self.client.batch.jobs.cancel(job_id=self.batch_id)
-
-    def delete_provider_batch(self):
-        if self.batch.status in ["QUEUED", "RUNNING"]:
-            self.cancel_provider_batch()
-        elif self.batch.status == "SUCCESS" and self.batch.output_file_id:
-            self.delete_provider_file(file_id=self.batch.output_file_id)
-
-    def get_provider_results(self) -> Response:
-        return self.client.files.download(file_id=self.batch.output_file)
-
-
-class GroqExperiment(Experiment):
-    body_cls: t.Type[GroqBody] = Field(
-        default=GroqBody, description="body class to use", init=False
-    )
-    request_cls: t.Type[GroqRequest] = Field(
-        default=GroqRequest, description="request class to use", init=False
-    )
-
-    @computed_field(repr=False)
-    @cached_property
-    def client(self) -> Groq:
-        """Get the client
-
-        Returns:
-            Groq: The client
-        """
-        return Groq(api_key=os.getenv(self.api_key_name))
-
-    def retrieve_provider_file(self):
-        return self.client.files.info(file_id=self.input_file_id)
-
-    def retrieve_provider_batch(self):
-        return self.client.batches.retrieve(batch_id=self.batch_id)
-
-    @computed_field
-    @property
-    def input_file(self) -> FileInfoResponse | None:
-        if self.input_file_id is None:
-            return None
-        return self.retrieve_provider_file()
-
-    @computed_field
-    @property
-    def batch(self) -> BatchRetrieveResponse | None:
-        if self.batch_id is None:
-            return None
-        return self.retrieve_provider_batch()
-
-    @computed_field
-    @property
-    def status(
-        self,
-    ) -> t.Literal[
-        "setup",
-        "created",
-        "validating",
-        "failed",
-        "in_progress",
-        "finalizing",
-        "completed",
-        "expired",
-        "cancelling",
-        "cancelled",
-    ]:
-        if self.batch_id is None:
-            if self.is_setup:
-                return "setup"
-            return "created"
-        return self.batch.status
-
-    def create_provider_file(self) -> str:
-        return self.client.files.create(file=open(self.input_file_path, "rb"), purpose="batch").id
-
-    def delete_provider_file(self, file_id: str):
-        self.client.files.delete(file_id=file_id)
-
-    def create_provider_batch(self) -> str:
-        return self.client.batches.create(
-            completion_window="24h",
-            endpoint=self.endpoint,
-            input_file_id=self.input_file_id,
-            metadata={"description": self.description},
-        ).id
-
-    def raise_not_in_running_status(self):
-        if self.status not in ["in_progress", "finalizing"]:
-            raise ValueError(
-                f"Experiment in status {self.status} is not in in_progress or finalizing status"
-            )
-
-    def raise_not_in_completed_status(self):
-        if self.status != "completed":
-            raise ValueError(f"Experiment in status {self.status} is not in completed status")
-
-    def cancel_provider_batch(self):
-        self.client.batches.cancel(batch_id=self.batch_id)
-
-    def delete_provider_batch(self):
-        if self.batch.status in ["in_progress", "finalizing"]:
-            self.cancel_provider_batch()
-        elif self.batch.status == "completed" and self.batch.output_file_id:
-            self.delete_provider_file(file_id=self.batch.output_file_id)
-
-    def get_provider_results(self) -> BinaryAPIResponse:
-        return self.client.files.content(file_id=self.batch.output_file_id)
